requirements.txt
camelot-py==0.10.1
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.8
pandas==1.3.3
pdfminer.six==20201018
pdftopng==0.2.3
pycparser==2.20
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sortedcontainers==2.4.0
tabulate==0.8.9
